Accept null return values in the loaders
This will hopefully fix problems with the android support that were
introduced with "Fix loader not loading certain functions". I don't
know if this is a problem with erupt or with the Android Loader
because as far as i know erupt will only request function pointers
which it is allowed to request.